requirements.txt
asgiref==3.7.2
boto3==1.28.57
botocore==1.31.57
certifi==2023.7.22
distlib==0.3.7
dj-database-url==2.1.0
Django==4.2.5
django-cors-headers==4.2.0
django-extensions==3.2.3
django-storages==1.14.1
djangorestframework==3.14.0
filelock==3.12.4
gunicorn==21.2.0
jmespath==1.0.1
packaging==23.1
Pillow==10.0.1
platformdirs==3.10.0
psycopg==3.1.12
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==1.26.16
virtualenv==20.24.5
